+++ packages/vendure-plugin-metrics/src/entities/metric-request.entity.ts
@@ -17,6 +17,15 @@ export class MetricRequest extends VendureEntity {
   deviceType!: string;
 
   @Index()
-  @Column()
-  channelToken!: string;
+  @Column({ type: 'varchar' })
+  channelId!: string | number;
+
+  @Column({ type: 'varchar', nullable: true })
+  path?: string;
+
+  @Column({ type: 'varchar', nullable: true })
+  productId?: string | number;
+
+  @Column({ type: 'varchar', nullable: true })
+  productVariantId?: string | number;
 }