+++ packages/vendure-plugin-metrics/src/index.ts
@@ -1,6 +1,6 @@
 export * from './metrics.plugin';
 export * from './services/metrics.service';
-export * from './api/metrics.resolver';
+export * from './api/metrics.admin-resolver';
 export * from './ui/generated/graphql';
 export * from './metrics/average-order-value';
 export * from './metrics/revenue-per-product';